initial event listing setup
