+++ src/components/events/Events.css
@@ -13,7 +13,7 @@
     line-height: 110%;
 }
 
-.ticketsub {
+.eventType {
     font-size: 13px;
     text-transform: uppercase;
     font-weight: 500;
